+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - DataHandler class to process Coinbase market and user data
+- Unit test for UserThreadWebsocketCallbacks multiple client support
 
 ### Changed
 - Try to find dependent slick components using `find_package` before falling back to `FetchContent`
 - Changed header files from .h to .hpp
 - Decoupled UserThreadWebsocketCallbacks from WebSocketeClient to support multiple WebSocketClient
+
+### Fixed
+- Various WebSocket unit tests not waiting for snapshot
 
 ## [0.1.0] - 2026-01-13